tutorials/ai_agents/07_receipts_agent/requirements.txt
openai==1.99.9
python-dotenv==1.1.0
pydantic==2.11.3
requests==2.32.3
vectorize-client==0.4.0